Add blob storage, header encoding, and compression for articles (WIP)
Refactored the code to add classes related to Article blob storage, specifically Article.cs and StorageEntry.cs. Correspondingly, a HeaderEncoding class was added to handle encoding and decoding of headers. XZ Compression package was added to the project as well, implemented in XZCompressor class, and its functionality was tested in XZCompressorTests. Class Article was renamed to ArticleHeader, reflecting its use more accurately.