Add article downloader and auxiliary classes
This commit introduces an `ArticleDownloader` class for downloading articles in chunks. It also includes the auxiliary classes and methods needed for this process such as `ArticleBlobFile` and `CustomNntpArticleResponse`. Changes in types and names, like `BlogOffset` from int to long in `ArticleHeader` and renaming DbSet property from `Article` to `ArticleHeader` in `GroupDbContext`, were also made correspondingly.