Refactor migrations and adjust date format
Migrations corresponding to old database structure have been removed and a fresh migration reflecting the current structure has been created. This change is coinciding with the switch from using DateTimeOffset to DateTime in the ArticleHeader entity. This update provides necessary adjustments to keep the project up-to-date with its database dependencies.